+++ src/renderer/src/components/message/CodeInline.tsx
@@ -3,5 +3,5 @@ type Props = {
 };
 
 export const CodeInline = ({ children }: Props) => {
-  return <span className="bg-gray-950 text-white rounded-sm px-1.5 py-0.5 text-xxs font-semibold">{children}</span>;
+  return <span className="bg-neutral-900 text-white rounded-sm px-1 py-0.5 text-xxs font-semibold">{children}</span>;
 };